requirements.txt
marqo
sentence-transformers>=2.2.0
python-dotenv
langchain
google-generativeai
redis